fix(pdf): made pdf work
